+++ src/main/java/org/sopt/www/Seminar/dto/member/MemberUpdateRequest.java
@@ -1,4 +1,4 @@
-package org.sopt.www.Seminar.dto.request.member;
+package org.sopt.www.Seminar.dto.member;
 
 import org.sopt.www.Seminar.domain.Part;
