+++ src/main/java/org/sopt/www/Seminar/dto/post/PostCreateRequest.java
@@ -1,4 +1,4 @@
-package org.sopt.www.Seminar.dto.request.post;
+package org.sopt.www.Seminar.dto.post;
 
 public record PostCreateRequest(String title, String content) {
