+++ src/main/java/org/sopt/www/Seminar/dto/post/PostGetResponse.java
@@ -1,4 +1,4 @@
-package org.sopt.www.Seminar.dto.request.post;
+package org.sopt.www.Seminar.dto.post;
 
 import org.sopt.www.Seminar.domain.Category;
 import org.sopt.www.Seminar.domain.Post;
