+++ src/main/java/org/sopt/www/Seminar/dto/post/PostUpdateRequest.java
@@ -1,4 +1,4 @@
-package org.sopt.www.Seminar.dto.request.post;
+package org.sopt.www.Seminar.dto.post;
 
 public record PostUpdateRequest(String content) {
 }
