+++ src/main/java/org/sopt/www/Seminar/service/CategoryService.java
@@ -4,7 +4,7 @@ import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.sopt.www.Seminar.domain.Category;
 import org.sopt.www.Seminar.domain.CategoryId;
-import org.sopt.www.Seminar.dto.request.category.CategoryResponse;
+import org.sopt.www.Seminar.dto.category.CategoryResponse;
 import org.sopt.www.Seminar.repository.CategoryJpaRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
