+++ src/main/java/org/sopt/www/Seminar/service/MemberService.java
@@ -3,9 +3,9 @@ package org.sopt.www.Seminar.service;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.sopt.www.Seminar.domain.Member;
-import org.sopt.www.Seminar.dto.request.member.MemberCreateRequest;
-import org.sopt.www.Seminar.dto.request.member.MemberUpdateRequest;
-import org.sopt.www.Seminar.dto.response.MemberGetResponse;
+import org.sopt.www.Seminar.dto.member.MemberCreateRequest;
+import org.sopt.www.Seminar.dto.member.MemberGetResponse;
+import org.sopt.www.Seminar.dto.member.MemberUpdateRequest;
 import org.sopt.www.Seminar.repository.MemberJpaRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
