+++ src/main/java/org/sopt/www/Seminar/service/PostService.java
@@ -4,9 +4,9 @@ import lombok.RequiredArgsConstructor;
 import org.sopt.www.Seminar.domain.Category;
 import org.sopt.www.Seminar.domain.Member;
 import org.sopt.www.Seminar.domain.Post;
-import org.sopt.www.Seminar.dto.request.post.PostCreateRequest;
-import org.sopt.www.Seminar.dto.request.post.PostGetResponse;
-import org.sopt.www.Seminar.dto.request.post.PostUpdateRequest;
+import org.sopt.www.Seminar.dto.post.PostCreateRequest;
+import org.sopt.www.Seminar.dto.post.PostGetResponse;
+import org.sopt.www.Seminar.dto.post.PostUpdateRequest;
 import org.sopt.www.Seminar.repository.MemberJpaRepository;
 import org.sopt.www.Seminar.repository.PostJpaRepository;
 import org.springframework.stereotype.Service;
